Use client for config network to get token metadata
